controle de acesso com gates pronto

+++ app/Http/Controllers/UserController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class UserController extends Controller
 {
@@ -51,6 +52,9 @@ class UserController extends Controller
      */
     public function edit(User $user)
     {
+        if (! Gate::allows('isAdm', $user)) {
+            return redirect()->route('users.index');
+        }
         return view('admin.users.edit', compact('user'));
     }
 
@@ -63,6 +67,9 @@ class UserController extends Controller
      */
     public function update(Request $request, User $user)
     {
+        if (! Gate::allows('isAdm', $user)) {
+            abort(403);
+        }
         $user->update($request->all());
         return redirect()->route('users.index');
     }
@@ -75,6 +82,9 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        if (! Gate::allows('isAdm', $user)) {
+            abort(403);
+        }
         User::destroy($user->id);
         return redirect()->route('users.index');
     }
